telegram_bot/requirements.txt
python-telegram-bot==10.0.2
SQLAlchemy==1.2.3
psycopg2==2.7.3